done stuff, need to make user input now

+++ C++ Assessment 2, Takeaway App/Beverage.cpp
@@ -1,12 +1,12 @@
 #include "Beverage.h"
 
 //constructor
-Beverage::Beverage(std::string name, double price, int calories, float abv, int volume)
+Beverage::Beverage(std::string name, double price, int calories, double abv, int volume)
 	:Item(name, price, calories), abv(abv), volume(volume) {}
 
 
 //get functions
-float Beverage::getAbv()
+double Beverage::getAbv()
 {
 	return abv;
 }
